Refactor engine structure to improve shader and camera management
The code changes mostly concern the mechanism of shader and camera creation and binding within the engine. A Shader and a Camera are now created and bound during the setup process at the Engine level. They are also passed onto OpenGLRenderData when created. This eliminates ad-hoc shader and camera update operations during the runtime, leading to cleaner and more maintainable code. These changes also allow easier modification of shader and camera settings within the app.