Add Unit Tests
This commit introduces settings for the OpenGLRenderer such as the existence of a platform, and adds corresponding tests. Also, the STRLEngine has been renamed to STRLFramework across CMakeLists.txt and code for more suitable components naming. This update advances the GLFW platform implementation and accommodates for window resizing and termination handling.