Update unit selection, movement, and formation system
This commit improves the selection mechanism by introducing selection rectangle for units, along with enhancing logic for unit movement. Furthermore, this update includes the introduction of a formation class, which is in charge of group unit movement. On the side, a bug in the destroy script found in the native script handler was successfully addressed and fixed.